requirements.txt
PyYAML==5.4.1
python-docx==0.8.10